pyproject.toml
[tool.poetry]
name = "segmentation"
version = "0.1.0"
description = ""
authors = ["hrsma2i <hrs.ma2i@gmail.com>"]
packages = [{ include = "segmentation" }]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
torch = { version = "1.6.0", optional = true }
torchvision = { version = "0.7.0", optional = true }
mmcv-full = { version = "^1.2.7", optional = true }
mmsegmentation = { version = "^0.11.0", optional = true }
typer = { version = "^0.3.2", optional = true }
numpy = { version = "^1.20.1", optional = true }
matplotlib = { version = "^3.3.4", optional = true }
pandas = "^1.2.3"

[tool.poetry.extras]
fastscnn = ["torch", "torchvision", "mmcv-full", "mmsegmentation", "typer"]
vis = ["numpy", "matplotlib"]

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
